feat(master): 引入Redission，使用读写锁控制清理临时图片定时任务的执行

+++ personal-blog-common/src/main/java/com/gewuyou/blog/common/vo/ArticleVO.java
@@ -24,7 +24,7 @@ import static com.gewuyou.blog.common.constant.MessageConstant.ARTICLE_TITLE_NOT
 @NoArgsConstructor
 @Schema(description = "文章")
 public class ArticleVO {
-    @Schema(description = "文章id", type = "integer", example = "1")
+    @Schema(description = "文章id", type = "Long", example = "1")
     private Long id;
 
     @NotEmpty(message = ARTICLE_TITLE_NOT_EMPTY)
